init sorting class
